Index related methods moved
